# server/docs/swagger.yaml
basePath: /
...
host: localhost:8090
info:
  contact: {}
  description: Minimalist torrent streaming server API.
  title: TorrServer API
  version: "1.0"
paths:
...
  /torrent/upload:
    post:
      consumes:
      - multipart/form-data
      description: Only one file support.
      parameters:
      - description: Torrent file to insert
        in: formData
        name: file
        required: true
        type: file
      - description: Save to DB
        in: formData
        name: save
        type: string
      - description: Torrent title
        in: formData
        name: title
        type: string
      - description: Torrent category
        in: formData
        name: category
        type: string
      - description: Torrent poster
        in: formData
        name: poster
        type: string
      - description: Torrent data
        in: formData
        name: data
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Torrent status
          schema:
            type: object
      summary: Add .torrent file
      tags:
      - API
...
schemes:
- http
swagger: "2.0"